INSP: move lint inspections and related classes to separate package
