+++ src/main/kotlin/org/rust/ide/inspections/lints/RsLintLevel.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.inspections
+package org.rust.ide.inspections.lints
 
 /**
  * Rust lint warning levels.
