+++ src/test/kotlin/org/rust/ide/inspections/lints/NamingNotationsTest.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.inspections
+package org.rust.ide.inspections.lints
 
 import junit.framework.TestCase
