+++ src/test/kotlin/org/rust/ide/inspections/lints/naming/RsEnumNamingInspectionTest.kt
@@ -3,10 +3,10 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.inspections.naming
+package org.rust.ide.inspections.lints.naming
 
-import org.rust.ide.inspections.RsEnumNamingInspection
 import org.rust.ide.inspections.RsInspectionsTestBase
+import org.rust.ide.inspections.lints.RsEnumNamingInspection
 
 class RsEnumNamingInspectionTest : RsInspectionsTestBase(RsEnumNamingInspection::class) {
     fun `test enums`() = checkByText("""
