+++ src/test/kotlin/org/rust/ide/inspections/lints/naming/RsMacroNamingInspectionTest.kt
@@ -3,10 +3,10 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.inspections.naming
+package org.rust.ide.inspections.lints.naming
 
 import org.rust.ide.inspections.RsInspectionsTestBase
-import org.rust.ide.inspections.RsMacroNamingInspection
+import org.rust.ide.inspections.lints.RsMacroNamingInspection
 
 class RsMacroNamingInspectionTest : RsInspectionsTestBase(RsMacroNamingInspection::class) {
     fun `test macros`() = checkByText("""
