+++ src/test/kotlin/org/rust/ide/inspections/lints/naming/RsMethodNamingInspectionTest.kt
@@ -3,10 +3,10 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.inspections.naming
+package org.rust.ide.inspections.lints.naming
 
 import org.rust.ide.inspections.RsInspectionsTestBase
-import org.rust.ide.inspections.RsMethodNamingInspection
+import org.rust.ide.inspections.lints.RsMethodNamingInspection
 
 class RsMethodNamingInspectionTest : RsInspectionsTestBase(RsMethodNamingInspection::class) {
     fun `test methods`() = checkByText("""
