+++ src/test/kotlin/org/rust/ide/inspections/lints/naming/RsTypeAliasNamingInspectionTest.kt
@@ -3,10 +3,10 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.inspections.naming
+package org.rust.ide.inspections.lints.naming
 
 import org.rust.ide.inspections.RsInspectionsTestBase
-import org.rust.ide.inspections.RsTypeAliasNamingInspection
+import org.rust.ide.inspections.lints.RsTypeAliasNamingInspection
 
 class RsTypeAliasNamingInspectionTest : RsInspectionsTestBase(RsTypeAliasNamingInspection::class) {
     fun `test type aliases`() = checkByText("""
